refactor: modularize backend and rename recovery functions
