+++ backend/src/lib/crypto.ts
