+++ backend/src/lib/errors.ts
