+++ backend/src/modules/admin/recovery/index.ts
@@ -1,0 +1,1 @@
+export * from "./service.ts";
